fix(widget): cast inner radius to double to fix analyzer

+++ studytrack/lib/core/widgets/neon_glow_container.dart
@@ -32,7 +32,7 @@ class NeonGlowContainer extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final radius = BorderRadius.circular(borderRadius);
-    final innerRadius = math.max(0, borderRadius - borderWidth);
+    final innerRadius = math.max(0, borderRadius - borderWidth).toDouble();
     final innerBorderRadius = BorderRadius.circular(innerRadius);
     final resolvedGradient = gradient ?? AppColors.primaryGradient;
